requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
billiard==4.1.0
black==23.7.0
cassandra-driver==3.28.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cqlengine==0.21.0
ecdsa==0.18.0
fastapi==0.101.1
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
kombu==5.3.1
mypy-extensions==1.0.0
orjson==3.9.5
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pydantic==2.2.1
pydantic-settings==2.0.3
pydantic_core==2.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
redis==5.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
